refactor providers directory
