+++ src/app/register/page.tsx
@@ -2,7 +2,6 @@
 
 import { QueryClient, QueryClientProvider } from 'react-query'
 import { Register } from '@/components/page/Register'
-import { UserDataProvider } from '../providers/UserDataProvider'
 
 const queryClient = new QueryClient()
