+++ src/components/page/Main/index.tsx
@@ -1,7 +1,7 @@
 import { useEffect } from 'react'
 import Image from 'next/image'
-import { useAuthContext } from '@/app/providers/AuthProvider'
-import { useUserDataContext } from '@/app/providers/UserDataProvider'
+import { useAuthContext } from '@/providers/AuthProvider'
+import { useUserDataContext } from '@/providers/UserDataProvider'
 
 const OWNER = '<OWNER>'
 const REPO = '<REPO>'
